+++ app/build.gradle.kts
@@ -77,7 +77,7 @@ dependencies {
     implementation("androidx.appcompat:appcompat:1.6.1")
     implementation("com.google.android.material:material:1.10.0")
     implementation("androidx.constraintlayout:constraintlayout:2.1.4")
-
+    implementation("androidx.test.ext:junit-ktx:1.1.5")
     implementation("androidx.cardview:cardview:1.0.0")
 
     //Lifacycle
@@ -151,12 +151,14 @@ dependencies {
 
     //Tests
     testImplementation("junit:junit:4.13.2")
-    androidTestImplementation("androidx.test:runner:1.5.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.5")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
+
+    androidTestImplementation("androidx.test:runner:1.5.2")
+    androidTestImplementation("androidx.test:core:1.5.0")
     testImplementation("io.mockk:mockk:1.12.4")
     testImplementation("com.squareup.okhttp3:mockwebserver:5.0.0-alpha.11")
-
+    testImplementation("android.arch.core:core-testing:1.1.1")
     testImplementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:1.7.1")
     androidTestImplementation("com.android.support:support-annotations:28.0.0")
     debugImplementation("androidx.fragment:fragment-testing:1.6.2")
@@ -178,6 +180,7 @@ dependencies {
     androidTestImplementation("com.google.dagger:hilt-android-testing:2.37")
     kaptTest("com.google.dagger:hilt-android-compiler:2.48.1")
     kaptAndroidTest("com.google.dagger:hilt-android-compiler:2.48.1")
+
 }
