+++ app/src/test/java/com/bonface/openweather/data/CurrentWeatherDaoTest.kt
@@ -1,4 +1,4 @@
-package com.bonface.openweather.data.local
+package com.bonface.openweather.data
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
